Latest Version

+++ paddle.h
@@ -17,15 +17,16 @@ class Paddle : public QGraphicsItem
   void paint(QPainter *painter, const QStyleOptionGraphicsItem *option,
              QWidget *widget);
 
+
  protected:
     void advance(int step);
     void keyPressEvent(QKeyEvent *event);
     void keyReleaseEvent(QKeyEvent *event);
 
  private:
-    int x;
-    int y;
-    int speed;
+    float x;
+    float y;
+    float speed;
     int input;
 };
